requirements.txt
ultralytics==8.0.196
deep-sort-realtime==1.3.2
opencv-python==4.8.1.78
shapely==2.0.2
PyYAML==6.0.1
tqdm==4.66.1

torch==2.0.1+cu118 --extra-index-url https://download.pytorch.org/whl/cu118

